Add --globs and replace V1 goal
