+++ src/python/pants/backend/project_info/register.py
@@ -5,7 +5,6 @@ from pants.backend.project_info.rules import source_file_validator
 from pants.backend.project_info.tasks.dependencies import Dependencies
 from pants.backend.project_info.tasks.depmap import Depmap
 from pants.backend.project_info.tasks.export import Export
-from pants.backend.project_info.tasks.filedeps import FileDeps
 from pants.backend.project_info.tasks.idea_plugin_gen import IdeaPluginGen
 from pants.goal.task_registrar import TaskRegistrar as task
 
@@ -20,7 +19,6 @@ def register_goals():
 
   task(name='depmap', action=Depmap).install()
   task(name='dependencies', action=Dependencies).install()
-  task(name='filedeps', action=FileDeps).install('filedeps')
 
 
 def rules():
