+++ src/python/pants/backend/project_info/tasks/BUILD
@@ -7,7 +7,6 @@ target(
     ':dependencies',
     ':depmap',
     ':export',
-    ':filedeps',
     ':idea_plugin_gen',
   ],
 )
@@ -57,18 +56,6 @@ python_library(
     'src/python/pants/java:executor',
     'src/python/pants/task',
     'src/python/pants/util:memo',
-  ],
-)
-
-python_library(
-  name = 'filedeps',
-  sources = ['filedeps.py'],
-  dependencies = [
-    'src/python/pants/backend/jvm/targets:jvm',
-    'src/python/pants/backend/jvm/targets:scala',
-    'src/python/pants/base:build_environment',
-    'src/python/pants/build_graph',
-    'src/python/pants/task',
   ],
 )
